+++ vedo/assembly.py
@@ -23,18 +23,6 @@ __all__ = ["Group", "Assembly", "procrustes_alignment"]
 
 
 #################################################
-def _is_sequence(arg):
-    """Check if the input is iterable."""
-    if hasattr(arg, "strip"):
-        return False
-    if hasattr(arg, "__getslice__"):
-        return True
-    if hasattr(arg, "__iter__"):
-        return True
-    return False
-
-
-#################################################
 def procrustes_alignment(sources, rigid=False):
     """
     Return an ``Assembly`` of aligned source meshes with the `Procrustes` algorithm.
@@ -57,13 +45,13 @@ def procrustes_alignment(sources, rigid=False):
         ![](https://vedo.embl.es/images/basic/align4.png)
     """
 
-    group = vtk.vtkMultiBlockDataGroupFilter()
+    group = vtk.get("MultiBlockDataGroupFilter")()
     for source in sources:
         if sources[0].npoints != source.npoints:
             vedo.logger.error("sources have different nr of points")
             raise RuntimeError()
         group.AddInputData(source.dataset)
-    procrustes = vtk.vtkProcrustesAlignmentFilter()
+    procrustes = vtk.get("ProcrustesAlignmentFilter")()
     procrustes.StartFromCentroidOn()
     procrustes.SetInputConnection(group.GetOutputPort())
     if rigid:
@@ -254,7 +242,7 @@ class Assembly(CommonVisual, vtk.vtkAssembly):
 
         scalarbars = []
         for a in self.actors:
-            if isinstance(a, vtk.vtkProp3D):  # and a.GetNumberOfPoints():
+            if isinstance(a, vtk.get("Prop3D")):  # and a.GetNumberOfPoints():
                 self.AddPart(a)
             if hasattr(a, "scalarbar") and a.scalarbar is not None:
                 scalarbars.append(a.scalarbar)
@@ -339,7 +327,7 @@ class Assembly(CommonVisual, vtk.vtkAssembly):
         """
         Add an object to the assembly
         """
-        if isinstance(obj, vtk.vtkProp3D):
+        if isinstance(obj, vtk.get("Prop3D")):
 
             self.objects.append(obj)
             self.actors.append(obj.actor)