+++ vedo/pyplot.py
@@ -3,10 +3,7 @@
 import warnings
 import numpy as np
 
-try:
-    import vedo.vtkclasses as vtk
-except ImportError:
-    import vtkmodules.all as vtk
+import vedo.vtkclasses as vtk
 
 import vedo
 from vedo import settings